fix lazy imports
